backend/requirements.txt
google-genai
streamlit
pandas
plotly
scikit-learn
python-dotenv
requests